requirements.txt
# ===== 猎影 (Shadow Hunt) v3.0 依赖库 =====
# 97工作室 - 2026-04-06

# ===== 核心框架 =====
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
python-multipart==0.0.9

# ===== 感知层 =====
supervision==0.19.0
ultralytics==8.1.0
deep-sort-realtime==1.3.2
opencv-contrib-python==4.9.0.80
av==14.0.0
ffmpeg-python==0.2.0

# ===== 认知层 =====
llama-cpp-python==0.2.40
transformers==4.38.0

# ===== 异步任务调度 =====
celery==5.3.6
redis==5.0.1
kombu==5.3.3

# ===== 向量检索 =====
faiss-cpu==1.7.4

# ===== Re-ID 特征提取 =====
torch==2.2.0
torchvision==0.17.0
torchreid==0.2.0

# ===== 认证与权限 =====
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ===== PDF 报告生成 =====
reportlab==4.0.8
weasyprint==60.2
pdfkit==1.0.0

# ===== 数据库 =====
sqlalchemy==2.0.25
alembic==1.13.1

# ===== 系统监控 =====
psutil==5.9.8
watchdog==4.0.0
py-cpuinfo==9.0.0

# ===== CPU 加速 =====
openvino==2024.0.0
onnxruntime==1.17.0

# ===== 工具库 =====
python-dotenv==1.0.1
pyyaml==6.0.1
pillow==10.2.0
numpy==1.26.4
pandas==2.2.0
tqdm==4.66.1
loguru==0.7.2
aiofiles==23.2.1
httpx==0.26.0
easydict==1.13
scipy==1.12.0
scikit-learn==1.4.0
requests==2.31.0
rich==13.7.0